+++ x/domain/types/types.go
@@ -68,6 +68,8 @@ func ValidateDomainType(typ DomainType) error {
 }
 
 // Account defines an account that belongs to a domain
+// NOTE: It should not be confused with cosmos-sdk auth account
+// github.com/cosmos/cosmos-sdk/x/auth.Account
 type Account struct {
 	// Domain references the domain this account belongs to
 	Domain string `json:"domain"`