+++ CHANGELOG.md
@@ -2,8 +2,14 @@
 
 ## HEAD
 
+## v0.4.5
+- CHANGE: bump cosmos-sdk and tendermint version
+- FIX: AccountRenewalCountMax and DomainRenewalCountMax bumped at configuration update
+- FIX: fix cli tests
 - FIX: fix domain renewal
-- BREAKING: account type field Name to pointer
+- Implement block-metrics
+- disable block metrics CI
+
 ## v0.4.4
 
 - fix: iovnscli get config